wstępna refaktoryzacja
