+++ dynamiczne/progress/WWY26100.csv
@@ -1,0 +1,4 @@
+Wyjazd,12/05/2023,0.0,1200.0
+Tankowanie,18/05/2023,380.0
+Powrót,20/05/2023,2900.0,420.0
+Norma,20/05/2023,40.0
