+++ dynamiczne/start_cli_proces.rb
