--- files/DW 12345
@@ -1,4 +1,0 @@
-wyjazd,2017-01-01
-tankowanie,2017-01-02
-tankowanie,2017-01-03
-powrot,2017-01-04
