--- progress/WWL21221.csv
@@ -1,1 +1,0 @@
-Wyjazd,12.12.12,0.0,1090.0
