+++ statyczne/cli.rb
