+++ statyczne/files/WWY38360.csv
@@ -1,0 +1,2 @@
+Data wyjazdu,Data powrotu,Dotankowane paliwo,Zużyte paliwo,Przejechane kilometry,Norma l/100km
+06/06/2023,15/06/2023,0,780.0,2900.0,26.89655172413793
